requirements.txt
absl-py==0.10.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
emoji==0.6.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
fsspec==0.8.4
future==0.18.2
google-auth==1.22.1
google-auth-oauthlib==0.4.1
grpcio==1.32.0
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Markdown==3.3.1
MarkupSafe==1.1.1
numpy==1.19.2
oauthlib==3.1.0
packaging==20.4
pandas==1.1.3
protobuf==3.13.0
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.0.2
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.15
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.3
sentencepiece==0.1.91
six==1.15.0
soynlp==0.0.493
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
tokenizers==0.8.1rc2
torch==1.6.0
tqdm==4.50.2
transformers==3.3.1
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.3.1
